requirements.txt
# VIPL Email Agent — Dependencies
google-api-python-client==2.114.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
anthropic==0.42.0
httpx==0.27.0
pyyaml==6.0.1
apscheduler==3.10.4
jinja2==3.1.3
python-dateutil==2.8.2
pytz==2024.1
tenacity==8.2.3
pymupdf==1.24.3
